build: add --auto-baseurl option
It is quite common to build a juliet website somewhere on the local
filesystem, for example, to check how some article renders. While
this is perfectly fine, the baseurl variable remains a problem: it
will almost certainly be different from the production baseurl.

In order to cope with this issue, users currently have to define a
whole new config.yml and use --config-file, or modify their
production config.yml.

This commit introduces a very simple solution to this problem: when
building on the local filesystem, simply pass --auto-baseurl. This
will automatically set the baseurl to the build destination folder
without the need to modify any config file.