build: rename slug variable to installed_filename
slug is kind of a bad variable, as it provides only part of the final
file name for posts and pages. Currently theme designers have to
append .html themeselves to get full installed file name. This adds
pointless complexity to themes and makes the code less modular, e.g.
if we want to modify installed extension.

Rename slug to installed_filename and makes sure it contains the
_real_ complete installed file name instead of the slug only.