new: rework theme header feature
Previous releases introduced a "theme config" feature.  This feature
allows themes to provide juliet with a list of page/post default
header entries.

This feature was badly named and did not work. This commit addresses
this issue for juliet new. Future commits will make sure that default
entries are also considered juliet build if a post does not specify
it.